CalendarApicopy/requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.3
django-google-auth==0.4.0
django-google-auth2==0.0.8
djangorestframework==3.14.0
freezegun==1.2.2
future==0.18.3
gkeepapi==0.14.2
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-api-keys==0.5.2
googleapis-common-protos==1.59.1
gpsoauth==1.0.2
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
pyglet==2.0.8
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
